# docs/swagger.yaml
basePath: /federation/v1
definitions:
...
  dto.OnboardApplicationRequest:
    properties:
      appId:
        type: string
      appName:
        type: string
      artefactId:
        type: string
    type: object
...
  models.Application:
    properties:
      appId:
        type: string
      appPkgId:
        description: is empty if the app instance is running on a partner zone
        type: string
      artefactId:
        type: string
      description:
        type: string
      federationContextId:
        type: string
      name:
        type: string
    type: object
...
  models.InvalidParam:
    properties:
      param:
        type: string
      reason:
        type: string
    type: object
...
  models.ProblemDetails:
    properties:
      cause:
        description: Fixed string indicating cause of the issue
        type: string
      detail:
        description: Specific detail of the issue
        type: string
      invalidParams:
        items:
          $ref: '#/definitions/models.InvalidParam'
        type: array
      title:
        description: Summary of the problem
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is the API documentation for the MEC Federator.
  title: MEC Federator API
  version: "1.0"
paths:
...
  /ewbi/{federationContextId}/applications:
    post:
      consumes:
      - application/json
      description: Onboards an application to the orchestrator and creates an application
        object in the federation
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Onboard Application Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OnboardApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Application onboarded successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Onboard Application
      tags:
      - EWBI - ApplicationOnboarding
  /ewbi/{federationContextId}/applications/{appId}:
    delete:
      consumes:
      - application/json
      description: Removes an application from the orchestrator and deletes the application
        object from the federation
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Application ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Application removed successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request or application has running instances
          schema:
            $ref: '#/definitions/models.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Remove Application
      tags:
      - EWBI - ApplicationOnboarding
    get:
      consumes:
      - application/json
      description: Retrieves application details by application ID within a federation
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Application ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Application'
        "400":
          description: Invalid request or application not found
          schema:
            $ref: '#/definitions/models.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: View Application
      tags:
      - EWBI - ApplicationOnboarding
    patch:
      consumes:
      - application/json
      description: Updates an application object in the federation
      parameters:
      - description: Federation Context ID
        in: path
        name: federationContextId
        required: true
        type: string
      - description: Application ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Application updated successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ProblemDetails'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ProblemDetails'
      summary: Update Application
      tags:
      - EWBI - ApplicationOnboarding
...
swagger: "2.0"